vms-430: widen the LLP64 syscall RETURN leg to 64-bit (Alpha mmap-backed malloc)
The return-leg counterpart of the already-applied syscall-ARGUMENT width fix
(__scc / syscall_arg_t == long long). On the alpha-dec-vms LLP64 port `long` is
32 bits while pointers and the Linux-Alpha syscall result register are 64 bits,
so the `long`-typed return path narrowed pointer-returning syscalls (mmap) to
their low 32 bits. Rail symptom: crtl_rms's mallocng (behind decc$_malloc64)
got buf=0xa000 from a truncated 0x2000_0000_a000 mmap address, then
memset(0xa000,...) SIGSEGV'd.

Widened sites (only the pointer-carrying + shared paths; non-pointer returns
like read/write byte counts are left as-is):
  - arch/alpha-dec-vms/syscall_arch.h: __syscall0..6 return long -> long long
    (keeps the backend __vms_alpha_syscall's 64-bit result intact).
  - src/internal/syscall_ret.c (new overlay): __syscall_ret now
    long long __syscall_ret(unsigned long long) — both legs 64-bit; the
    -4096 errno sentinel/mapping preserved.
  - src/internal/syscall.h decl + src/mman/mmap.c `long ret` local: widened via
    idempotent, post-state-guarded sed in build-musl.sh (stock musl is pinned +
    checksum-verified). syscall_ret.c is wired into the overlay copy list.

mremap/sbrk/shmat/lite_malloc/mallocng carry pointers through _Addr(=long long)
types or route through __syscall_ret, so they are fixed by the two shared
changes with no per-file edit. __syscall_cp (non-pointer, cancellable path) is
deliberately left `long`.

Proof: FULL alpha-dec-vms libc.a (1346 members) rebuilds with the widening — no
over-widening breakage. Codegen probe: the OLD `long` return emits
`addl $31,$0,$0` (sign-extend-from-longword = truncate $0 to 32 bits) on both a
raw and a pointer-cast mmap return; the NEW `long long` return removes it,
carrying the full quadword. The Alpha rail verifies crtl_rms activates + runs to
completion (N=7) with the fix.

+++ tools/cross-alpha-vms/musl-arch/arch/alpha-dec-vms/syscall_arch.h
@@ -46,33 +46,48 @@ long long __vms_alpha_syscall(long long n, long long a1, long long a2,
 #define __SYSCALL_LL_E(x) (x)
 #define __SYSCALL_LL_O(x) (x)
 
-static inline long __syscall0(long long n)
+/*
+ * vms-430 -- 64-bit syscall RETURN ABI, the return-leg counterpart of the
+ * 64-bit ARGUMENT ABI above (__scc / syscall_arg_t == long long).
+ *
+ * On this LLP64 port `long` is only 32 bits, but the Linux-Alpha kernel returns
+ * a full 64-bit register: pointer-returning syscalls (mmap/mremap/brk) hand back
+ * a 64-bit address. A `long`-typed __syscallN would narrow that address to its
+ * low 32 bits (the compiler emits stl/ldl/sextl), so a genuine mmap result like
+ * 0x2000_0000_a000 collapses to 0xa000 -- and every mmap-backed malloc (musl
+ * mallocng behind decc$_malloc64) then memsets a bogus low address and SIGSEGVs.
+ * The backend __vms_alpha_syscall() already returns `long long`; keep that width
+ * all the way out by typing the raw return here as `long long` too. (The shared
+ * __syscall_ret is widened to match in src/internal/syscall_ret.c + syscall.h.)
+ * The argument side (__scc/syscall_arg_t) is unchanged.
+ */
+static inline long long __syscall0(long long n)
 {
 	return __vms_alpha_syscall(n, 0, 0, 0, 0, 0, 0);
 }
-static inline long __syscall1(long long n, long long a)
+static inline long long __syscall1(long long n, long long a)
 {
 	return __vms_alpha_syscall(n, a, 0, 0, 0, 0, 0);
 }
-static inline long __syscall2(long long n, long long a, long long b)
+static inline long long __syscall2(long long n, long long a, long long b)
 {
 	return __vms_alpha_syscall(n, a, b, 0, 0, 0, 0);
 }
-static inline long __syscall3(long long n, long long a, long long b, long long c)
+static inline long long __syscall3(long long n, long long a, long long b, long long c)
 {
 	return __vms_alpha_syscall(n, a, b, c, 0, 0, 0);
 }
-static inline long __syscall4(long long n, long long a, long long b, long long c,
+static inline long long __syscall4(long long n, long long a, long long b, long long c,
 			     long long d)
 {
 	return __vms_alpha_syscall(n, a, b, c, d, 0, 0);
 }
-static inline long __syscall5(long long n, long long a, long long b, long long c,
+static inline long long __syscall5(long long n, long long a, long long b, long long c,
 			     long long d, long long e)
 {
 	return __vms_alpha_syscall(n, a, b, c, d, e, 0);
 }
-static inline long __syscall6(long long n, long long a, long long b, long long c,
+static inline long long __syscall6(long long n, long long a, long long b, long long c,
 			     long long d, long long e, long long f)
 {
 	return __vms_alpha_syscall(n, a, b, c, d, e, f);
